+++ src/helpers/index.ts
@@ -3,7 +3,7 @@ import { config as Config } from '../utils'
 import { generate as generateEvaluationContext } from './evaluationContext'
 import { EvaluationContext } from './logicEvaluate'
 import { createStandardLibraryResolver } from './hardcodedMapping'
-import { Reporter, defaultReporter } from './reporter'
+import { Reporter, defaultReporter } from '../utils/reporter'
 import { FSWrapper, createFSWrapper } from './fs'
 export { HardcodedMap } from './hardcodedMapping'
 export { EvaluationContext } from './logicEvaluate'
